Small fixes to mission boss bars

+++ API/src/main/java/com/bgsoftware/superiorskyblock/api/missions/Mission.java
@@ -328,7 +328,7 @@ public abstract class Mission<V> {
      * Send boss bar progress
      */
     public void sendBossBar(SuperiorPlayer superiorPlayer, String action, double progress, double total, double totalProgress) {
-        missionCategory.sendBossBar(superiorPlayer, this, action, progress, total, totalProgress);
+        missionCategory.sendBossBar(superiorPlayer, this, action, (int) progress, (int) total, totalProgress);
     }
 
     @Override
